[thin_shrink] Pull out thin-related functions into own file
